Updated Architecture
Updated readme with steps to follow and updated the architecture for easier access
