+++ C++/Set 1/q1.cpp
