+++ C++/Set 1/q2.cpp
