+++ C++/Set 1/q3.cpp
