+++ C++/Set 2/q1.cpp
