+++ C++/Set 2/q2.cpp
