+++ C++/Set 2/q3.cpp
