+++ C++/Set 3/q1.cpp
