+++ C++/Set 3/q2.cpp
