+++ C++/Set 4/q1.cpp
