+++ C++/Set 4/q2.cpp
