+++ C++/Set 5/q1.cpp
