+++ C++/Set 5/q2.cpp
