+++ Python/linearSearch.py
