+++ Python/min_max_in_array_solution.py
