+++ Python/pallindrome_num.py
