+++ Python/perfect_num.py
