+++ Python/strong_num.py
