+++ Python/text_handwi.py
@@ -1,0 +1,9 @@
+import pywhatkit as pk
+pk.text_to_handwriting('''Learning Python from the basics is 
+extremely important. Before starting to learn python,understanding a 
+base language like c is a must and 
+some of the oops concepts.Python program has many modulesand packages, 
+which helps with efficient programming.
+Understanding these modules and 1proper usage of many syntax and libraries is recommended.
+In this article, a few modules and packages are used in the program. 
+Python includes tons of libraries and some of them are quiet intresting''')
