+++ Python/texttoaudio.py
@@ -1,0 +1,4 @@
+import pyttsx3
+engine=pyttsx3.init()
+engine.say("Hello Rashi , What your plan for tommorow??")
+engine.runAndWait()
