+++ README.md
@@ -1,2 +1,7 @@
-# Python
-this repository consits of python scripts and some DSA based questions which have been uploaded as a part of hacktober fest 2021. Anyone can contribute in this.
+ # THE CODE HUB
+This Repository is one stop destination to contribute your work and ideas and category.<br>
+
+## Steps
+- Each language has been assigned its own folder which you can fork and contribute.
+- If you want to contribute in any other language create its separate folder and contribute under it by forking it.
+- If you are solving a question from a coding platform please paste the link to the question in the starting of code in comment.
