--- texttoaudio.py
@@ -1,4 +1,0 @@
-import pyttsx3
-engine=pyttsx3.init()
-engine.say("Hello Rashi , What your plan for tommorow??")
-engine.runAndWait()
